Cleaned up the oort package and subpackages
