+++ oort/groupconfig_GEN_.go
@@ -1,4 +1,4 @@
-package api
+package oort
 
 import (
 	"crypto/rand"
@@ -12,7 +12,7 @@ import (
 	"google.golang.org/grpc"
 )
 
-// GroupStoreConfig defines the settings when calling NewGroupStore.
+// GroupStoreConfig defines the settings when creating new stores, such as NewReplGroupStore.
 type GroupStoreConfig struct {
 	// Logger defines where log output will go. If not set, the logger will
 	// default to zap.New(zap.NewJSONEncoder()).
