+++ oort/grouperrorstore_GEN_.go
@@ -1,4 +1,4 @@
-package api
+package oort
 
 import (
 	"fmt"
