+++ oort/replstore.got
@@ -1,4 +1,4 @@
-package api
+package oort
 
 import (
     "errors"
@@ -198,7 +198,7 @@ func (rs *Repl{{.T}}Store) storesFor(ctx context.Context, keyA uint64) ([]store.
             if ss[i] == nil {
                 ss[i] = rs.stores[as[i]]
                 if ss[i] == nil {
-                    ss[i] = NewPool{{.T}}Store(as[i], rs.poolSize, rs.concurrentRequestsPerStore, rs.ftlsConfig,  rs.grpcOpts...)
+                    ss[i] = newPool{{.T}}Store(as[i], rs.poolSize, rs.concurrentRequestsPerStore, rs.ftlsConfig,  rs.grpcOpts...)
                     rs.stores[as[i]] = ss[i]
                     select {
                     case <-ctx.Done():
