+++ oort/replstore_test.got
@@ -1,4 +1,4 @@
-package api
+package oort
 
 import (
     "testing"
