+++ oort/valuereplstore_GEN_test.go
@@ -1,4 +1,4 @@
-package api
+package oort
 
 import (
 	"testing"
